Control.Monad.Log

Description

MonadLog is a general purpose logging effect. Unlike other logging libraries available on Hackage, MonadLog does not assume that you will be logging text information. Instead, the choice of logging data is up to you. This, combined with mapLogMessage and the ability to reintepret log messages results in an extremely powerful set of logging combinators.

This library also comes with a flexible default implementation of MonadLog. By using Handlers along with their Contravariant, Decidable and Divisible instances, you are able to route log messages to multiple sources.

We'll explore this power in the tutorial.

Tutorial:

To add logging to your applications, you will need to make two changes.

First, use the MonadLog type class to indicate that a computation has access to logging. MonadLog is parameterized on the type of messages that you intend to log. In this example, we will log Text that is wrapped in the WithUrgency. Note that this does not specify how to perform the actual logging.

testApp :: MonadLog (WithUrgency Text) m => m ()
testApp = do
  logMessage (WithUrgency Info "Don't mind me")
  logMessage (WithUrgency Error "But do mind me!")

Next, we need to run this computation under a MonadLog effect handler. In this example, we open two file descriptor loggers, so we can dispatch our log events to STDOUT and STDERR. We use choose to split out log messages between these two Handlers - messages with a msgUrgency of Error will go to STDERR, and all other messages will go to STDOUT.

Note that both our STDOUT and STDERR handlers only know how to log Text (see the type of withFDHandler). However, we're logging WithUrgency Text, so we need to adjust these handles. To do so, we use contramap to adjust their inputs. For stdoutHandler, I use renderWithUrgency to render the urgency information, and then pass the log message through unchanged. For stderrHandler I also render the urgency information, but I also transform the text into all caps - I really don't want to miss those messages!

withSplitLogging :: LoggingT (WithUrgency Text) IO () -> IO ()
withSplitLogging m =
  withFDHandler stderr $ stderrHandler ->
  withFDHandler stdout $ stdoutHandler ->
  runLoggingT m
              (choose (message ->
                        case msgUrgency message of
                          Error -> Left message
                          _ -> Right message)
                      (contramap (renderWithUrgency id) stdoutHandler)
                      (contramap (renderWithUrgency (T.map toUpper)) stdoutHandler))

MonadLog

class Monad m => MonadLog message m | m -> message where

The class of monads that support logging.

Methods

logMessage :: message -> m ()

Append a message to the log for this computation.

mapLogMessage :: MonadLog message' m => (message -> message') -> LoggingT message m a -> m a

Re-interpret the log messages in one computation. This can be useful to embed a computation with one log type in a larger general computation.

LoggingT, a general handler

data LoggingT message m a

LoggingT is a very general handler for the MonadLog effect. Whenever a log entry is emitted, the given Handler is invoked, producing some side-effect (such as writing to stdout, or appending a database table).

pattern LoggingT :: (Handler t t -> t t) -> LoggingT t t t

runLoggingT :: LoggingT message m a -> Handler m message -> m a

data Handler m message

Handlers are mechanisms to interpret the meaning of logging as an action in the underlying monad. The instances of Contravariant, Decidable and Divisible can all be used to adapt Handlers so they can be used in different contexts.

pattern Handler :: (b -> m ()) -> Handler m b

withFDHandler :: MonadIO io => Handle -> (Handler io Text -> io a) -> io a

newFDHandler creates a new Handler that will append a given file descriptor (or Handle, as it is known in the "base" library). Note that this Handler requires log messages to be of type Text. You can use contramap, divide and choose to adapt Handlers appropriately.

These Handlers asynchronously log messages to the given file descriptor, rather than blocking.

Message wrappers

newtype WithTimestamp a

Constructors

WithTimestamp

Fields

unWithTimestamp :: (UTCTime, a)

withTimestamps :: (MonadLog (WithTimestamp message) m, MonadIO m) => LoggingT message m a -> m a

Add timestamps to all messages logged. Timestamps will be calculated synchronously when log entries are emitted.

renderWithUrgency :: (a -> Text) -> WithUrgency a -> Text

data WithUrgency a

Add "urgency" information to a log message. This is often used to convey how significant a log message is.

Constructors

WithUrgency

Fields

msgUrgency :: Urgency

discardUrgency :: a

data Urgency

Constructors

Info

Debug

Warning

Error

renderWithTimestamp :: (a -> Text) -> WithTimestamp a -> Text