Control.Monad.Log

Description

MonadLog is a general purpose logging effect. Unlike other logging libraries available on Hackage, MonadLog does not assume that you will be logging text information. Instead, the choice of logging data is up to you. This, combined with mapLogMessage and the ability to reintepret log messages results in an extremely powerful set of logging combinators.

This library also comes with a flexible default implementation of MonadLog.

We'll explore this power in the tutorial.

Tutorial:

To add logging to your applications, you will need to make two changes.

First, use the MonadLog type class to indicate that a computation has access to logging. MonadLog is parameterized on the type of messages that you intend to log. In this example, we will log Text that is wrapped in the WithSeverity. Note that this does not specify how to perform the actual logging.

testApp :: MonadLog (WithSeverity Text) m => m ()
testApp = do
  logMessage (WithSeverity Info "Don't mind me")
  logMessage (WithSeverity Error "But do mind me!")

Next, we need to run this computation under a MonadLog effect handler. In this example, we open two file descriptor loggers, so we can dispatch our log events to STDOUT and STDERR. We use choose to split out log messages between these two Handlers - messages with a msgSeverity of Error will go to STDERR, and all other messages will go to STDOUT.

Note that both our STDOUT and STDERR handlers only know how to log Text (see the type of withFDHandler). However, we're logging WithSeverity Text, so we need to adjust these handles. To do so, we use function pre-composition to adjust their inputs. For stdoutHandler, I use renderWithSeverity to render the severity information, and then pass the log message through unchanged. For stderrHandler I also render the severity information, but I also transform the text into all caps - I really don't want to miss those messages!

withSplitLogging :: LoggingT (WithSeverity Text) IO () -> IO ()
withSplitLogging m =
  withFDHandler defaultBatchingOptions stderr $ stderrHandler ->
  withFDHandler defaultBatchingOptions stdout $ stdoutHandler ->
  runLoggingT m 
              (\message ->
                 case msgSeverity message of
                   Error -> stderrHandler (renderWithSeverity id message)
                   _ ->
                     stdoutHandler
                       (renderWithSeverity (T.map toUpper)
                                           message))

MonadLog

class Monad m => MonadLog message m | m -> message where

The class of monads that support logging.

Methods

logMessage :: message -> m ()

Append a message to the log for this computation.

mapLogMessage :: MonadLog message' m => (message -> message') -> LoggingT message m a -> m a

Re-interpret the log messages in one computation. This can be useful to embed a computation with one log type in a larger general computation.

Message wrappers

Timestamps

data WithTimestamp a

Add a timestamp to log messages.

Constructors

WithTimestamp

Fields

discardTimestamp :: a

Retireve the time a message was logged.

msgTimestamp :: UTCTime

View the underlying message.

withTimestamps :: (MonadLog (WithTimestamp message) m, MonadIO m) => LoggingT message m a -> m a

Add timestamps to all messages logged. Timestamps will be calculated synchronously when log entries are emitted.

renderWithTimestamp

Arguments

:: (UTCTime -> String)
How to format the timestamp.

-> (a -> Text)
How to render the rest of the message.

-> WithTimestamp a -> Text

Given a way to render the underlying message a and a way to format UTCTime, render a message with its timestamp.

>>> renderWithTimestamp (formatTime defaultTimeLocale rfc822DateFormat) id timestamppedLogMessage
[Tue, 19 Jan 2016 11:29:42 UTC] Setting target speed to plaid

Severity

data WithSeverity a

Add "Severity" information to a log message. This is often used to convey how significant a log message is.

Constructors

WithSeverity

Fields

msgSeverity :: Severity

Retrieve the Severity a message.

discardSeverity :: a

View the underlying message.

data Severity

Classes of severity for log messages. These have been chosen to match syslog severity levels

Constructors

Emergency
System is unusable. By syslog convention, this level should not be used by applications.

Alert
Should be corrected immediately.

Critical
Critical conditions.

Error
Error conditions.

Warning
May indicate that an error will occur if action is not taken.

Notice
Events that are unusual, but not error conditions.

Informational
Normal operational messages that require no action.

Debug
Information useful to developers for debugging the application.

renderWithSeverity :: (a -> Text) -> WithSeverity a -> Text

Given a way to render the underlying message a render a message with its timestamp.

>>> renderWithSeverity id Debug (WithSeverity Info "Flux capacitor is functional")
[Info] Flux capacitor is functional

LoggingT, a general handler

data LoggingT message m a

LoggingT is a very general handler for the MonadLog effect. Whenever a log entry is emitted, the given Handler is invoked, producing some side-effect (such as writing to stdout, or appending a database table).

pattern LoggingT :: (Handler t t -> t t) -> LoggingT t t t

LoggingT messasge m is isomorphic to Handler message m -> m (). This is a reader monad with base effects in m and access to Handler message m.This pattern synonym witnesses that isomorphism.

runLoggingT :: LoggingT message m a -> Handler m message -> m a

Given a Handler for a given message, interleave this Handler into the underlying m computation whenever logMessage is called.

mapLoggingT

Arguments

:: (Handler n message' -> Handler m message)
Given a handler for the new monad n, embed this handler inside the original monad m.

-> (forall a. m a -> n a)
A natural transformation/monad morphism from the original monad m to the new monad n. For example, you could use liftIO or runWriterT.

-> LoggingT message m a

-> LoggingT message' n a

LoggingT unfortunately does admit an instance of the MFunctor type class, which provides the hoist method to change the monad underneith a monad transformer. However, it is possible to do this with LoggingT provided that you have a way to re-interpret a log handler in the original monad.

LoggingT Handlers

type Handler m message = message -> m ()

Handlers are mechanisms to interpret the meaning of logging as an action in the underlying monad. They are simply functions from log messages to m-actions.

withFDHandler :: (MonadIO io, MonadMask io) => BatchingOptions -> Handle -> (Handler io Text -> io a) -> io a

withFDHandler creates a new Handler that will append a given file descriptor (or Handle, as it is known in the "base" library). Note that this Handler requires log messages to be of type Text.

These Handlers asynchronously log messages to the given file descriptor, rather than blocking.

Batched handlers

withBatchedHandler :: (MonadIO io, MonadMask io) => BatchingOptions -> ([message] -> IO ()) -> (Handler io message -> io a) -> io a

Create a new batched handler. Batched handlers take batches of messages to log at once, which can be more performant than logging each individual message.

A batched handler flushes under three criteria:

The flush interval has elapsed and the queue is not empty.

The queue has become full and needs to be flushed.

The scope of withBatchedHandler is exited.

Batched handlers queue size and flush period can be configured via BatchingOptions.

data BatchingOptions

Options that be used to configure withBatchingHandler.

Constructors

BatchingOptions

Fields

flushMaxDelay :: Int

The maximum amount of time to wait between flushes

flushMaxQueueSize :: Int

The maximum amount of messages to hold in memory between flushes}

defaultBatchingOptions :: BatchingOptions

Defaults for BatchingOptions

defaultBatchingOptions = BatchingOptions { flushMaxDelay = 1000000, flushMaxQueueSize = 100 }

Pure logging

newtype PureLoggingT log m a

A MonadLog handler optimised for pure usage. Log messages are accumulated strictly, given that messasges form a Monoid.

Constructors

MkPureLoggingT (StateT log m a)

runPureLoggingT :: Monoid log => PureLoggingT log m a -> m (a, log)

Discarding logs

data DiscardLoggingT message m a

discardLogging :: DiscardLoggingT message m a -> m a